requirements.txt
fastapi
uvicorn[standard]
pydantic
pymysql
selenium
undetected-chromedriver
beautifulsoup4
requests
python-dotenv
